+++ packages/eyes-appium/package.json
@@ -47,8 +47,7 @@
     "selenium-webdriver": "^4.0.0-alpha.5"
   },
   "scripts": {
-    "link": "node ../../scripts/link.js eyes-appium",
-    "unlink": "node ../../scripts/unlink.js eyes-appium",
+    "install": "yarn install --focused",
     "lint": "npx eslint . --ext .js --no-eslintrc --config ../../.eslintrc",
     "test": "mocha --no-timeouts \"test/**/*.spec.js\"",
     "test:it": "mocha --no-timeouts \"test/it/**/*.spec.js\"",
