+++ packages/eyes-images-browser/package.json
@@ -38,8 +38,7 @@
     "uglify-es": "^3.3.9"
   },
   "scripts": {
-    "link": "node ../../scripts/link.js eyes-images-browser",
-    "unlink": "node ../../scripts/unlink.js eyes-images-browser",
+    "install": "yarn install --focused",
     "lint": "npx eslint . --ext .js --no-eslintrc --config ../../.eslintrc",
     "browserify": "browserify index.js > dist/eyes-images.browser.js",
     "uglify": "uglifyjs dist/eyes-images.browser.js > dist/eyes-images.browser.min.js",
