+++ packages/eyes-protractor/package.json
@@ -41,8 +41,7 @@
     "protractor": "^6.0.0-beta"
   },
   "scripts": {
-    "link": "node ../../scripts/link.js eyes-protractor",
-    "unlink": "node ../../scripts/unlink.js eyes-protractor",
+    "install": "yarn install --focused",
     "lint": "npx eslint . --ext .js --no-eslintrc --config ../../.eslintrc",
     "test": "protractor test/protractor.conf.js",
     "test:it": "protractor test/protractor.conf.js"
