+++ packages/eyes-sdk-core-browser/package.json
@@ -38,8 +38,7 @@
     "uglify-es": "^3.3.9"
   },
   "scripts": {
-    "link": "node ../../scripts/link.js eyes-sdk-core-browser",
-    "unlink": "node ../../scripts/unlink.js eyes-sdk-core-browser",
+    "install": "yarn install --focused",
     "lint": "npx eslint . --ext .js --no-eslintrc --config ../../.eslintrc",
     "browserify": "browserify index.js > dist/eyes-sdk-core.browser.js",
     "uglify": "uglifyjs dist/eyes-sdk-core.browser.js > dist/eyes-sdk-core.browser.min.js",
