+++ packages/eyes-selenium/package.json
@@ -57,8 +57,7 @@
     "selenium-webdriver": "^4.0.0-alpha.5"
   },
   "scripts": {
-    "link": "node ../../scripts/link.js eyes-selenium",
-    "unlink": "node ../../scripts/unlink.js eyes-selenium",
+    "install": "yarn install --focused",
     "lint": "npx eslint . --ext .js --no-eslintrc --config ../../.eslintrc",
     "test": "mocha --no-timeouts \"test/**/*.spec.js\"",
     "test:unit": "mocha --no-timeouts \"test/unit/**/*.spec.js\"",
