+++ packages/sdk-test-kit/package.json
@@ -19,8 +19,7 @@
     "url": "git+https://github.com/applitools/eyes.sdk.javascript1.git"
   },
   "scripts": {
-    "link": "node ../../scripts/link.js sdk-test-kit",
-    "unlink": "node ../../scripts/unlink.js sdk-test-kit",
+    "install": "yarn install --focused",
     "lint": "npx eslint . --ext .js --no-eslintrc --config ../../.eslintrc",
     "test": "mocha test/**/*.spec.js",
     "build": "tsc src/coverage-tests/api.ts"
